backend/requirements.txt
datasets
symspellpy
indic-nlp-library
hunspell
protobuf
accelerate
certifi
charset-normalizer
filelock
fsspec
huggingface-hub
idna
Jinja2
MarkupSafe
mpmath
networkx
numpy
packaging
peft
pillow
psutil
PyYAML
regex
requests
safetensors
sentencepiece
sympy
tokenizers
torch
torchaudio
torchvision
tqdm
transformers
typing-extensions
urllib3
fastapi
uvicorn
pyjwt
firebase-admin
python-multipart
dotenv
